renamed files navItems for clarity
